0.2.0
This is the second release of daily. I've been using it long enough now
to feel safe removing the "alpha" suffix.

This release should be backwards-compatible with the previous version.
This release adds batch editing for entries, gives daily the ability to
display its own version, and fixes a copule of bugs.