Lock parsedatetime dep to 2.5
The author of this library has made some volatile changes as of late
(within the past year)

Version1  has the flag system where a flag of 0 indicates failure to
parse, 1 is date, 2 is time, and 3 is datetime. Ok that makes sense

Then Version 2 added a ctx system. Ok fine, but then it also started
printing a warning message even if the function was explicitly provided
the arugment to perform the "flag" behavior.

Now the ctx is default, and the message is gone, but using a version
older than 2 (eg 1.5) will incorrectly parse dates formatted by daily.
Like "2022-08-27, Sat" Will result in "2022-09-03, Sat" instead. No idea
why.

So I'm locking the version down. To the author of this lib; please stop
fucking with the interface I need your shit to work.