DAILY_12: Better attributes
Also addresses DAILY_9

Attributes are now displayed in a pretty way when the rst or markdown is
rendered and the fields are no longer hardcoded. Instead the attributes
are now in yaml format

The text after the special line "<!--- attributes --->" is parsed as a
yaml dict. Same for rst, but for the string ".. code-block:: yaml"

The attributes are also indented so they will be rendered as a code
block in md and rst.

This will require modifying existing journal entries, though this
release will already represent a breaking change from 0.4.0 .