DAILY_4: Individual files for entries
Each journal entry will now be stored as its own file. This is to add
compatibility with vimwiki.

The other features are preserved; per-heading display / editing, batch
editing of entries, etc...

Also added unit tests for date parsing and the journal class.

Added no_edit function. This can be used to generate a blank entry
without editing. This option can be used with vimwiki templates to
generate new entries with vim.

Conversion
----------
This commit is a breaking change. To convert your journal from the .json
file to the new files first a dailiy show command to show all journal
entries in a raw text format and redirect that output to a file matching
the date convention of YYYY-mm-dd.[rst,md]. When you run daily add while
directing the journal to this directory it should write each entry to
its own file.