DAILY_5: Markdown Support
Added support for markdown journals.

Not *super* clean, given that the journal being read should probably be
the entity that contains the config option, but w/e since we changing
over to individual files after this anyway to support vimwiki.

But this is technically backwards compatible, if clunky. So that's a
small gift to any who prefer the monolithic json file. If they want to
continue then we can fork from this point.

Also adds unit tests.