DAILY_16: Removed debian packaging support.
Packaging for debian was too complicated, especially when compared to
just using `pip3 --user`

I'll just tag and push those to the source repo. Clone and pip is the
current installation method.

Also fixes the installation dirs of the data files to relative pathing.
This will require the user to properly set up their bash autocomplete
and MANPATH environment variables.