feature: --copy-previous
This adds the ability to copy a previous entry's content when creating a
new entry.

This should be useful for keeping track of information like monthly
goals, weekly todos, etc... Stuff that doesn't change often but you want
to be reminded of it.

Also you get to yesterday's entry in front of you so you can see the
changes immediately.