fix: Removed copy-previous from config
Calling it a fix and not a remove; that's how annoying the config file
was to use with copy-previous. Bad idea.

Note to self; better usability testing in the future.

That said, a copy_previous in a daily.ini will still be accepted. The
following statement should be regarded as a coincidence of
implementation, but any of daily's command line arguments may have
defaults set in the config file.